# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cash Track API",
    "version": "1.0.0",
    "description": "API de gestión de movimientos, usuarios y reportes para la prueba técnica."
  },
  "paths": {
...
    "/api/movements": {
      "get": {
        "summary": "Listar movimientos financieros",
        "description": "Devuelve una lista paginada de movimientos (ingresos y egresos). Los administradores ven todos los movimientos, los usuarios solo los propios.",
        "tags": [
          "Movements"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Número de página para la paginación."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Cantidad de elementos por página."
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de movimientos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "concept": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "number"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "INCOME",
                              "EXPENSE"
                            ]
                          },
                          "date": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "user": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string",
                                "nullable": true
                              },
                              "email": {
                                "type": "string",
                                "format": "email"
                              },
                              "phone": {
                                "type": "string",
                                "nullable": true
                              },
                              "role": {
                                "type": "string",
                                "enum": [
                                  "USER",
                                  "ADMIN"
                                ]
                              }
                            }
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo movimiento",
        "description": "Crea un nuevo ingreso o egreso asociado al usuario autenticado.",
        "tags": [
          "Movements"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "concept",
                  "amount",
                  "date",
                  "type"
                ],
                "properties": {
                  "concept": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "INCOME",
                      "EXPENSE"
                    ]
                  }
                }
              },
              "example": {
                "concept": "Pago de salario",
                "amount": 1500.5,
                "date": "2025-01-15T00:00:00.000Z",
                "type": "INCOME"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Movimiento creado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "concept": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "number"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "INCOME",
                        "EXPENSE"
                      ]
                    },
                    "date": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos en la solicitud."
          },
          "401": {
            "description": "No autenticado."
          },
          "403": {
            "description": "No autorizado, se requiere rol ADMIN."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        }
      }
    },
...
  },
...
}